requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.40.0
webauthn>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
